now with seasonal limits

+++ src/Deluxxe/Raffles/JsonRaffleResultWriter.cs
@@ -1,21 +1,23 @@
+using Deluxxe.IO;
 using Microsoft.Extensions.Logging;
 
 namespace Deluxxe.Raffles;
 
 using System.Text.Json;
 
-public class JsonRaffleResultWriter(ILogger<JsonRaffleResultWriter> logger, string outputDirectory, bool overWrite = false) : IRaffleResultWriter
+public class JsonRaffleResultWriter(ILogger<JsonRaffleResultWriter> logger, JsonRaffleSerializerOptions options) : IRaffleResultWriter
 {
     private readonly JsonSerializerOptions _options = new()
     {
         WriteIndented = true,
     };
 
-    public async Task<bool> WriteAsync(RaffleResult result, CancellationToken cancellationToken = default)
+    public async Task<Uri> WriteAsync(RaffleResult result, CancellationToken cancellationToken = default)
     {
-        var file = new FileInfo(Path.Combine(Path.GetFullPath(outputDirectory), $"{result.season}-{result.name}-results.json"));
+        var fileName = $"{result.season}-{result.name}-results.json";
+        var file = new FileInfo(Path.Combine(Path.GetFullPath(options.outputDirectory), fileName));
 
-        if (file.Exists)
+        if (file.Exists && options.shouldOverwrite)
         {
             logger.LogInformation($"deleting previous file at {file.FullName}");
             file.Delete();
@@ -25,6 +27,6 @@ public class JsonRaffleResultWriter(ILogger<JsonRaffleResultWriter> logger, stri
         await using Stream stream = new FileStream(file.FullName, FileMode.OpenOrCreate);
         await JsonSerializer.SerializeAsync(stream, result, cancellationToken: cancellationToken, options: _options);
         stream.Close();
-        return true;
+        return FileUriParser.Generate(options.outputDirectory, fileName);
     }
 }
